requirements.txt
pandas~=2.2.3
requests~=2.32.3
python-dotenv